chmullighw class based on HMC homework template
Designed to be easy to use and provide good, clean headers.
It can be used with the HMC Math department boxes, or can
be used with a more traditional \begin{enumerate} style.
By default includes a long header on the front page, short
header on each page, page numbers at bottom center.
That's configurable with noheader and norunheader options.